Fix Diag & Solve loading for ASE
